Bhattacharjee {extraDistr}	R Documentation

Bhattacharjee distribution

Description

Density, distribution function, and random generation for the Bhattacharjee distribution.

Usage

dbhatt(x, mu = 0, sigma = 1, a = sigma, log = FALSE)

pbhatt(q, mu = 0, sigma = 1, a = sigma, lower.tail = TRUE,
  log.p = FALSE)

rbhatt(n, mu = 0, sigma = 1, a = sigma)

Arguments

x, q
vector of quantiles.

mu, sigma, a
location, scale and shape parameters. Scale and shape must be positive.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

If Z ~ Normal(0, 1) and U ~ Uniform(0, 1), then Z+U follows Bhattacharjee distribution.

Probability density function

f(z) = 1/(2*a) * (Φ((x-μ+a)/σ) - Φ((x-μ+a)/σ))

Cumulative distribution function

F(z) = σ/(2*a) * ((x-μ)*Φ((x-μ+a)/σ) - (x-μ)*Φ((x-μ-a)/σ) + φ((x-μ+a)/σ) - φ((x-μ-a)/σ))

References

Bhattacharjee, G.P., Pandit, S.N.N., and Mohan, R. (1963). Dimensional chains involving rectangular and normal error-distributions. Technometrics, 5, 404-406.

Examples

x <- rbhatt(1e5, 5, 3, 5)
hist(x, 100, freq = FALSE)
curve(dbhatt(x, 5, 3, 5), -20, 20, col = "red", add = TRUE)
hist(pbhatt(x, 5, 3, 5))
plot(ecdf(x))
curve(pbhatt(x, 5, 3, 5), -20, 20, col = "red", lwd = 2, add = TRUE)

[Package extraDistr version 1.8.6 Index]