BirnbaumSaunders {extraDistr}	R Documentation

Birnbaum-Saunders (fatigue life) distribution

Description

Density, distribution function, quantile function and random generation for the Birnbaum-Saunders (fatigue life) distribution.

Usage

dfatigue(x, alpha, beta = 1, mu = 0, log = FALSE)

pfatigue(q, alpha, beta = 1, mu = 0, lower.tail = TRUE, log.p = FALSE)

qfatigue(p, alpha, beta = 1, mu = 0, lower.tail = TRUE, log.p = FALSE)

rfatigue(n, alpha, beta = 1, mu = 0)

Arguments

x, q
vector of quantiles.

alpha, beta, mu
shape, scale and location parameters. Scale and shape must be positive.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability density function

f(x) = ((sqrt((x-μ)/β) + sqrt(β/(x-μ)))/(2*α*(x-μ))) * φ((sqrt((x-μ)/β) - sqrt(β/(x-μ)))/α)

Cumulative distribution function

F(x) = Φ(((sqrt((x-μ)/β) - sqrt(β/(x-μ)))/α)

Quantile function

F^-1(p) = (α/2 * Φ^-1(p) + sqrt((α/2 * Φ^-1(p))^2 + 1)^2 * β + μ

References

Birnbaum, Z. W. and Saunders, S. C. (1969). A new family of life distributions. Journal of Applied Probability, 6(2), 637-652.

Desmond, A. (1985) Stochastic models of failure in random environments. Canadian Journal of Statistics, 13, 171-183.

Vilca-Labra, F., and Leiva-Sanchez, V. (2006). A new fatigue life model based on the family of skew-elliptical distributions. Communications in Statistics-Theory and Methods, 35(2), 229-244.

Leiva, V., Sanhueza, A., Sen, P. K., and Paula, G. A. (2008). Random number generators for the generalized Birnbaum-Saunders distribution. Journal of Statistical Computation and Simulation, 78(11), 1105-1118.

Examples

x <- rfatigue(1e5, .5, 2, 5)
hist(x, 100, freq = FALSE)
curve(dfatigue(x, .5, 2, 5), 2, 20, col = "red", add = TRUE)
hist(pfatigue(x, .5, 2, 5))
plot(ecdf(x))
curve(pfatigue(x, .5, 2, 5), 2, 20, col = "red", lwd = 2, add = TRUE)

[Package extraDistr version 1.8.6 Index]