BivPoiss {extraDistr}	R Documentation

Bivariate Poisson distribution

Description

Probability mass function and random generation for the bivariate Poisson distribution.

Usage

dbvpois(x, y = NULL, a, b, c, log = FALSE)

rbvpois(n, a, b, c)

Arguments

x, y
vectors of quantiles; alternativelly x may be a two-column matrix (or data.frame) and y may be omitted.

a, b, c
positive valued parameters.

log
logical; if TRUE, probabilities p are given as log(p).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability mass function

f(x) = exp(-(a+b+c)) * (a^x)/x! * (b^y)/y! * sum(choose(x,k)*choose(y,k)*k!*(c/(a*b))^k)

References

Karlis, D. and Ntzoufras, I. (2003). Analysis of sports data by using bivariate Poisson models. Journal of the Royal Statistical Society: Series D (The Statistician), 52(3), 381-393.

Kocherlakota, S. and Kocherlakota, K. (1992) Bivariate Discrete Distributions. New York: Dekker.

Johnson, N., Kotz, S. and Balakrishnan, N. (1997). Discrete Multivariate Distributions. New York: Wiley.

Holgate, P. (1964). Estimation for the bivariate Poisson distribution. Biometrika, 51(1-2), 241-287.

Kawamura, K. (1984). Direct calculation of maximum likelihood estimator for the bivariate Poisson distribution. Kodai mathematical journal, 7(2), 211-221.

See Also

Poisson

Examples

x <- rbvpois(5000, 7, 8, 5)
image(prop.table(table(x[,1], x[,2])))
colMeans(x)

[Package extraDistr version 1.8.6 Index]