DirMnom {extraDistr}	R Documentation

Dirichlet-multinomial (multivariate Polya) distribution

Description

Density function, cumulative distribution function and random generation for the Dirichlet-multinomial (multivariate Polya) distribution.

Usage

ddirmnom(x, size, alpha, log = FALSE)

rdirmnom(n, size, alpha)

Arguments

x
k-column matrix of quantiles.

size
numeric vector; number of trials (zero or more).

alpha
k-values vector or k-column matrix; concentration parameter. Must be positive.

log
logical; if TRUE, probabilities p are given as log(p).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

If (p[1],…,p[k]) ~ Dirichlet(α[1],…,α[k]) and (x[1],…,x[k]) ~ Multinomial(n, p[1],…,p[k]), then (x[1],…,x[k]) ~ DirichletMultinomial(n, α[1],…,α[k]).

Probability density function

f(x) = (n! * Γ(sum(α[k]))) / (Γ(n + sum(α[k]))) * prod((Γ(x[k] + α[k])) / (x[k]! * Γ(α[k]))

References

Gentle, J.E. (2006). Random number generation and Monte Carlo methods. Springer.

Kvam, P. and Day, D. (2001) The multivariate Polya distribution in combat modeling. Naval Research Logistics, 48, 1-17.

See Also

Dirichlet, Multinomial

[Package extraDistr version 1.8.6 Index]