Dirichlet {extraDistr}	R Documentation

Dirichlet distribution

Description

Density function, cumulative distribution function and random generation for the Dirichlet distribution.

Usage

ddirichlet(x, alpha, log = FALSE)

rdirichlet(n, alpha)

Arguments

x
k-column matrix of quantiles.

alpha
k-values vector or k-column matrix; concentration parameter. Must be positive.

log
logical; if TRUE, probabilities p are given as log(p).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability density function

f(x) = Γ(sum(α[k])) / prod(Γ(α[k])) * prod(x[k]^{k-1})

References

Devroye, L. (1986). Non-Uniform Random Variate Generation. Springer-Verlag.

Examples

# Generating 10 random draws from Dirichlet distribution
# parametrized using a vector

rdirichlet(10, c(1, 1, 1, 1))

# or parametrized using a matrix where each row
# is a vector of parameters

alpha <- matrix(c(1, 1, 1, 1:3, 7:9), ncol = 3, byrow = TRUE)
rdirichlet(10, alpha)

[Package extraDistr version 1.8.6 Index]