DiscreteLaplace {extraDistr}	R Documentation

Discrete Laplace distribution

Description

Probability mass, distribution function and random generation for the discrete Laplace distribution parametrized by location and scale.

Usage

ddlaplace(x, location, scale, log = FALSE)

pdlaplace(q, location, scale, lower.tail = TRUE, log.p = FALSE)

rdlaplace(n, location, scale)

Arguments

x, q
vector of quantiles.

location
location parameter.

scale
scale parameter; 0 < scale < 1.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

If U ~ Geometric(1-p) and V ~ Geometric(1-p), then U-V ~ DiscreteLaplace(p), where geometric distribution is related to discrete Laplace distribution in similar way as exponential distribution is related to Laplace distribution.

Probability mass function

f(x) = (1-p)/(1+p) * p^(|x-μ|)

Cumulative distribution function

F(x) = [if x < 0:] p^-floor(x-μ))/(1+p) [else:] 1-(p^(floor(x-μ)+1))/(1+p)

References

Inusah, S., & Kozubowski, T.J. (2006). A discrete analogue of the Laplace distribution. Journal of statistical planning and inference, 136(3), 1090-1102.

Kotz, S., Kozubowski, T., & Podgorski, K. (2012). The Laplace distribution and generalizations: a revisit with applications to communications, economics, engineering, and finance. Springer Science & Business Media.

Examples

p <- 0.45
x <- rdlaplace(1e5, 0, p)
xx <- seq(-200, 200, by = 1)
plot(prop.table(table(x)))
lines(xx, ddlaplace(xx, 0, p), col = "red")
hist(pdlaplace(x, 0, p))
plot(ecdf(x))
lines(xx, pdlaplace(xx, 0, p), col = "red", type = "s")

[Package extraDistr version 1.8.6 Index]