DiscreteNormal {extraDistr}	R Documentation

Discrete normal distribution

Description

Probability mass function, distribution function and random generation for discrete normal distribution.

Usage

ddnorm(x, mean = 0, sd = 1, log = FALSE)

pdnorm(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

rdnorm(n, mean = 0, sd = 1)

Arguments

x, q
vector of quantiles.

mean
vector of means.

sd
vector of standard deviations.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability mass function

f(x) = Φ((x-μ+1)/σ) - Φ((x-μ)/σ)

Cumulative distribution function

F(x) = Φ((floor(x)+1-μ)/σ)

References

Roy, D. (2003). The discrete normal distribution. Communications in Statistics-Theory and Methods, 32, 1871-1883.

See Also

Normal

Examples

x <- rdnorm(1e5, 0, 3)
xx <- -15:15
plot(prop.table(table(x)))
lines(xx, ddnorm(xx, 0, 3), col = "red")
hist(pdnorm(x, 0, 3))
plot(ecdf(x))
xx <- seq(-15, 15, 0.1)
lines(xx, pdnorm(xx, 0, 3), col = "red", lwd = 2, type = "s")

[Package extraDistr version 1.8.6 Index]