DiscreteWeibull {extraDistr}	R Documentation

Discrete Weibull distribution (type I)

Description

Density, distribution function, quantile function and random generation for the discrete Weibull (type I) distribution.

Usage

ddweibull(x, shape1, shape2, log = FALSE)

pdweibull(q, shape1, shape2, lower.tail = TRUE, log.p = FALSE)

qdweibull(p, shape1, shape2, lower.tail = TRUE, log.p = FALSE)

rdweibull(n, shape1, shape2)

Arguments

x, q
vector of quantiles.

shape1, shape2
parameters (named q, β). Values of shape2 need to be positive and 0 < shape1 < 1.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability mass function

f(x) = q^x^β - q^(x+1)^β

Cumulative distribution function

F(x) = 1-q^(x+1)^β

Quantile function

F^-1(p) = ceiling((log(1-p)/log(q))^(1/β) - 1)

References

Nakagawa, T. and Osaki, S. (1975). The Discrete Weibull Distribution. IEEE Transactions on Reliability, R-24, 300-301.

Kulasekera, K.B. (1994). Approximate MLE's of the parameters of a discrete Weibull distribution with type I censored data. Microelectronics Reliability, 34(7), 1185-1188.

Khan, M.A., Khalique, A. and Abouammoh, A.M. (1989). On estimating parameters in a discrete Weibull distribution. IEEE Transactions on Reliability, 38(3), 348-350.

See Also

Weibull

Examples

x <- rdweibull(1e5, 0.32, 1)
xx <- seq(-2, 100, by = 1)
plot(prop.table(table(x)), type = "h")
lines(xx, ddweibull(xx, .32, 1), col = "red")

# Notice: distribution of F(X) is far from uniform:
hist(pdweibull(x, .32, 1), 50)

plot(ecdf(x))
lines(xx, pdweibull(xx, .32, 1), col = "red", lwd = 2, type = "s")

[Package extraDistr version 1.8.6 Index]