Frechet {extraDistr}	R Documentation

Frechet distribution

Description

Density, distribution function, quantile function and random generation for the Frechet distribution.

Usage

dfrechet(x, lambda = 1, mu = 0, sigma = 1, log = FALSE)

pfrechet(q, lambda = 1, mu = 0, sigma = 1, lower.tail = TRUE,
  log.p = FALSE)

qfrechet(p, lambda = 1, mu = 0, sigma = 1, lower.tail = TRUE,
  log.p = FALSE)

rfrechet(n, lambda = 1, mu = 0, sigma = 1)

Arguments

x, q
vector of quantiles.

lambda, sigma, mu
shape, scale, and location parameters. Scale and shape must be positive.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability density function

f(x) = λ/σ * ((x-μ)/σ)^(-1-λ) * exp(-((x-μ)/σ)^-λ)

Cumulative distribution function

F(x) = exp(-((x-μ)/σ)^-λ)

Quantile function

F^-1(p) = μ + σ * -log(p)^{-1/λ}

References

Bury, K. (1999). Statistical Distributions in Engineering. Cambridge University Press.

Examples

x <- rfrechet(1e5, 5, 2, 1.5)
xx <- seq(0, 1000, by = 0.1)
hist(x, 200, freq = FALSE)
lines(xx, dfrechet(xx, 5, 2, 1.5), col = "red") 
hist(pfrechet(x, 5, 2, 1.5))
plot(ecdf(x))
lines(xx, pfrechet(xx, 5, 2, 1.5), col = "red", lwd = 2)

[Package extraDistr version 1.8.6 Index]