GEV {extraDistr}	R Documentation

Generalized extreme value distribution

Description

Density, distribution function, quantile function and random generation for the generalized extreme value distribution.

Usage

dgev(x, mu = 0, sigma = 1, xi = 0, log = FALSE)

pgev(q, mu = 0, sigma = 1, xi = 0, lower.tail = TRUE, log.p = FALSE)

qgev(p, mu = 0, sigma = 1, xi = 0, lower.tail = TRUE, log.p = FALSE)

rgev(n, mu = 0, sigma = 1, xi = 0)

Arguments

x, q
vector of quantiles.

mu, sigma, xi
location, scale, and shape parameters. Scale must be positive.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability density function

f(x) = [if ξ != 0:] 1/σ * (1-ξ*(x-μ)/σ)^{-1-1/ξ} * exp(-(1-ξ*(x-μ)/σ)^{-1/ξ}) [else:] 1/σ * exp(-(x-μ)/σ) * exp(-exp(-(x-μ)/σ))

Cumulative distribution function

F(x) = [if ξ != 0:] exp(-(1+ξ*(x-μ)/σ)^{1/ξ}) [else:] exp(-exp(-(x-μ)/σ))

Quantile function

F^-1(p) = [if ξ != 0:] μ - σ/ξ * (1 - (-log(p))^ξ) [else:] μ - σ * log(-log(p))

References

Coles, S. (2001). An Introduction to Statistical Modeling of Extreme Values. Springer.

Examples

x <- rgev(1e5, 5, 2, .5)
hist(x, 1000, freq = FALSE, xlim = c(0, 50))
curve(dgev(x, 5, 2, .5), 0, 50, col = "red", add = TRUE)
hist(pgev(x, 5, 2, .5))
plot(ecdf(x))
curve(pgev(x, 5, 2, .5), 0, 50, col = "red", lwd = 2, add = TRUE)

[Package extraDistr version 1.8.6 Index]