Gumbel {extraDistr}	R Documentation

Gumbel distribution

Description

Density, distribution function, quantile function and random generation for the Gumbel distribution.

Usage

dgumbel(x, mu = 0, sigma = 1, log = FALSE)

pgumbel(q, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qgumbel(p, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rgumbel(n, mu = 0, sigma = 1)

Arguments

x, q
vector of quantiles.

mu, sigma
location and scale parameters. Scale must be positive.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability density function

f(x) = 1/σ * exp(-((x-μ)/σ + exp(-(x-μ)/σ)))

Cumulative distribution function

F(x) = exp(-exp(-(x-μ)/σ))

Quantile function

F^-1(p) = μ - σ * log(-log(p))

References

Bury, K. (1999). Statistical Distributions in Engineering. Cambridge University Press.

Examples

x <- rgumbel(1e5, 5, 2)
hist(x, 100, freq = FALSE)
curve(dgumbel(x, 5, 2), 0, 25, col = "red", add = TRUE)
hist(pgumbel(x, 5, 2))
plot(ecdf(x))
curve(pgumbel(x, 5, 2), 0, 25, col = "red", lwd = 2, add = TRUE)

[Package extraDistr version 1.8.6 Index]