InvGamma {extraDistr}	R Documentation

Inverse-gamma distribution

Description

Density, distribution function and random generation for the inverse-gamma distribution.

Usage

dinvgamma(x, alpha, beta = 1, log = FALSE)

pinvgamma(q, alpha, beta = 1, lower.tail = TRUE, log.p = FALSE)

qinvgamma(p, alpha, beta = 1, lower.tail = TRUE, log.p = FALSE)

rinvgamma(n, alpha, beta = 1)

Arguments

x, q
vector of quantiles.

alpha, beta
positive valued shape and scale parameters.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability mass function

f(x) = (x^(-α-1) * exp(-1/(β*x))) / (Γ(α)*β^α)

Cumulative distribution function

F(x) = γ(α, 1/(β*x)) / Γ(α)

References

Witkovsky, V. (2001). Computing the distribution of a linear combination of inverted gamma variables. Kybernetika 37(1), 79-90.

Leemis, L.M. and McQueston, L.T. (2008). Univariate Distribution Relationships. American Statistician 62(1): 45-53.

See Also

GammaDist

Examples

x <- rinvgamma(1e5, 20, 3)
hist(x, 100, freq = FALSE)
curve(dinvgamma(x, 20, 3), 0, 1, col = "red", add = TRUE)
hist(pinvgamma(x, 20, 3))
plot(ecdf(x))
curve(pinvgamma(x, 20, 3), 0, 1, col = "red", lwd = 2, add = TRUE) 

[Package extraDistr version 1.8.6 Index]