Laplace {extraDistr}	R Documentation

Laplace distribution

Description

Density, distribution function, quantile function and random generation for the Laplace distribution.

Usage

dlaplace(x, mu = 0, sigma = 1, log = FALSE)

plaplace(q, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qlaplace(p, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rlaplace(n, mu = 0, sigma = 1)

Arguments

x, q
vector of quantiles.

mu, sigma
location and scale parameters. Scale must be positive.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability density function

f(x) = 1/(2*σ) * exp(-|(x-μ)/σ|)

Cumulative distribution function

F(x) = [if x < mu:] 1/2 * exp((x-μ)/σ) [else:] 1 - 1/2 * exp((x-μ)/σ)

Quantile function

F^-1(p) = [if p < 0.5:] μ + σ * log(2*p) [else:] μ - σ * log(2*(1-p))

References

Krishnamoorthy, K. (2006). Handbook of Statistical Distributions with Applications. Chapman & Hall/CRC

Forbes, C., Evans, M. Hastings, N., & Peacock, B. (2011). Statistical Distributions. John Wiley & Sons.

Examples

x <- rlaplace(1e5, 5, 16)
hist(x, 100, freq = FALSE)
curve(dlaplace(x, 5, 16), -200, 200, n = 500, col = "red", add = TRUE)
hist(plaplace(x, 5, 16))
plot(ecdf(x))
curve(plaplace(x, 5, 16), -200, 200, n = 500, col = "red", lwd = 2, add = TRUE)

[Package extraDistr version 1.8.6 Index]