NSBeta {extraDistr}	R Documentation

Non-standard beta distribution

Description

Non-standard form of beta distribution with lower and upper bounds denoted as min and max. By default min=0 and max=1 what leads to standard beta distribution.

Usage

dnsbeta(x, shape1, shape2, min = 0, max = 1, log = FALSE)

pnsbeta(q, shape1, shape2, min = 0, max = 1, lower.tail = TRUE,
  log.p = FALSE)

qnsbeta(p, shape1, shape2, min = 0, max = 1, lower.tail = TRUE,
  log.p = FALSE)

rnsbeta(n, shape1, shape2, min = 0, max = 1)

Arguments

x, q
vector of quantiles.

shape1, shape2
non-negative parameters of the Beta distribution.

min, max
lower and upper bounds.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤q x], otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

See Also

Beta

Examples

x <- rnsbeta(1e5, 5, 13, -4, 8)
hist(x, 100, freq = FALSE)
curve(dnsbeta(x, 5, 13, -4, 8), -4, 6, col = "red", add = TRUE) 
hist(pnsbeta(x, 5, 13, -4, 8))
plot(ecdf(x))
curve(pnsbeta(x, 5, 13, -4, 8), -4, 6, col = "red", lwd = 2, add = TRUE)

[Package extraDistr version 1.8.6 Index]