NonStandardT {extraDistr}	R Documentation

Non-standard t-distribution

Description

Probability mass function, distribution function and random generation for non-standard t-distribution. Non-standard t-distribution besides degrees of freedom ν, is parametrized using additional parameters μ for location and σ for scale (μ=0 and σ = 1 for standard t-distribution).

Usage

dnst(x, df, mu = 0, sigma = 1, log = FALSE)

pnst(q, df, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qnst(p, df, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rnst(n, df, mu = 0, sigma = 1)

Arguments

x, q
vector of quantiles.

df
degrees of freedom (> 0, maybe non-integer). df = Inf is allowed.

mu
vector of locations

sigma
vector of positive valued scale parameters.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

See Also

TDist

Examples

x <- rnst(1e5, 1000, 5, 13)
hist(x, 100, freq = FALSE)
curve(dnst(x, 1000, 5, 13), -60, 60, col = "red", add = TRUE)
hist(pnst(x, 1000, 5, 13))
plot(ecdf(x))
curve(pnst(x, 1000, 5, 13), -60, 60, col = "red", lwd = 2, add = TRUE)

[Package extraDistr version 1.8.6 Index]