PoissonMix {extraDistr}	R Documentation

Mixture of Poisson distributions

Description

Density, distribution function and random generation for the mixture of Poisson distributions.

Usage

dmixpois(x, lambda, alpha, log = FALSE)

pmixpois(q, lambda, alpha, lower.tail = TRUE, log.p = FALSE)

rmixpois(n, lambda, alpha)

Arguments

x, q
vector of quantiles.

lambda
matrix (or vector) of (non-negative) means.

alpha
matrix (or vector) of mixing proportions; mixing proportions need to sum up to 1.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

n
number of observations. If length(n) > 1, the length is taken to be the number required.

p
vector of probabilities.

Details

Probability density function

f(x) = α[1] * f1(x; λ[1]) + … + α[k] * fk(x; λ[k])

Cumulative distribution function

F(x) = α[1] * F1(x; λ[1]) + … + α[k] * Fk(x; λ[k])

where sum(α[i]) == 1.

Examples

x <- rmixpois(1e5, c(5, 12, 19), c(1/3, 1/3, 1/3))
xx <- seq(-1, 50)
plot(prop.table(table(x)))
lines(xx, dmixpois(xx, c(5, 12, 19), c(1/3, 1/3, 1/3)), col = "red")
hist(pmixpois(x, c(5, 12, 19), c(1/3, 1/3, 1/3)))

xx <- seq(0, 50, by = 0.01)
plot(ecdf(x))
lines(xx, pmixpois(xx, c(5, 12, 19), c(1/3, 1/3, 1/3)), col = "red", lwd = 2)

[Package extraDistr version 1.8.6 Index]