Rayleigh {extraDistr}	R Documentation

Rayleigh distribution

Description

Density, distribution function, quantile function and random generation for the Rayleigh distribution.

Usage

drayleigh(x, sigma = 1, log = FALSE)

prayleigh(q, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qrayleigh(p, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rrayleigh(n, sigma = 1)

Arguments

x, q
vector of quantiles.

sigma
positive valued parameter.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability density function

f(x) = x/σ^2 * exp(-(x^2 / 2*σ^2))

Cumulative distribution function

F(x) = 1 - exp(-x^2 / 2*σ^2)

Quantile function

F^-1(p) = sqrt(-2*σ^2 * log(1-p))

References

Krishnamoorthy, K. (2006). Handbook of Statistical Distributions with Applications. Chapman & Hall/CRC.

Forbes, C., Evans, M. Hastings, N., & Peacock, B. (2011). Statistical Distributions. John Wiley & Sons.

Examples

x <- rrayleigh(1e5, 13)
hist(x, 100, freq = FALSE)
curve(drayleigh(x, 13), 0, 60, col = "red", add = TRUE)
hist(prayleigh(x, 13)) 
plot(ecdf(x))
curve(prayleigh(x, 13), 0, 60, col = "red", lwd = 2, add = TRUE) 

[Package extraDistr version 1.8.6 Index]