Skellam {extraDistr}	R Documentation

Skellam distribution

Description

Probability mass function and random generation for the Skellam distribution.

Usage

dskellam(x, mu1, mu2, log = FALSE)

rskellam(n, mu1, mu2)

Arguments

x
vector of quantiles.

mu1, mu2
positive valued parameters.

log
logical; if TRUE, probabilities p are given as log(p).

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

If X and Y follow Poisson distributions with means μ[1] and μ[2], than X-Y follows Skellam distribution parametrized by μ[1] and μ[2].

Probability mass function

f(x) = exp(-(μ1+μ2)) * (μ1/μ2)^(x/2) * besselI(2*sqrt(μ1*μ2), x)

References

Karlis, D., & Ntzoufras, I. (2006). Bayesian analysis of the differences of count data. Statistics in medicine, 25(11), 1885-1905.

Skellam, J.G. (1946). The frequency distribution of the difference between two Poisson variates belonging to different populations. Journal of the Royal Statistical Society, series A, 109(3), 26.

Examples

x <- rskellam(1e5, 5, 13)
xx <- -40:40
plot(prop.table(table(x)), type = "h")
lines(xx, dskellam(xx, 5, 13), col = "red")

[Package extraDistr version 1.8.6 Index]