ZIP {extraDistr}	R Documentation

Zero-inflated Poisson distribution

Description

Probability mass function and random generation for the zero-inflated Poisson distribution.

Usage

dzip(x, lambda, pi, log = FALSE)

pzip(q, lambda, pi, lower.tail = TRUE, log.p = FALSE)

qzip(p, lambda, pi, lower.tail = TRUE, log.p = FALSE)

rzip(n, lambda, pi)

Arguments

x, q
vector of quantiles.

lambda
vector of (non-negative) means.

pi
probability of extra zeros.

log, log.p
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].

p
vector of probabilities.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

Details

Probability density function

f(x) = [if x = 0:] (1-π)+π * exp(-λ) [else:] (1-π) * dpois(x, lambda)

See Also

Poisson

Examples

x <- rzip(1e5, 6, 0.33)
xx <- -2:20
plot(prop.table(table(x)), type = "h")
lines(xx, dzip(xx, 6, 0.33), col = "red")

xx <- seq(0, 20, by = 0.01)
plot(ecdf(x))
lines(xx, pzip(xx, 6, 0.33), col = "red")

[Package extraDistr version 1.8.6 Index]