extraDistr-package {extraDistr}	R Documentation

extraDistr: Additional Univariate and Multivariate Distributions

Description

Density, distribution function, quantile function and random generation for a number of univariate and multivariate distributions. This package implements the following distributions: Bernoulli, beta-binomial, beta-negative binomial, beta prime, Bhattacharjee, Birnbaum-Saunders, bivariate normal, bivariate Poisson, categorical, Dirichlet, Dirichlet-multinomial, discrete gamma, discrete Laplace, discrete normal, discrete uniform, discrete Weibull, Frechet, gamma-Poisson, generalized extreme value, Gompertz, generalized Pareto, Gumbel, half-Cauchy, half-normal, half-t, Huber density, inverse chi-squared, inverse-gamma, Kumaraswamy, Laplace, location-scale t, logarithmic, Lomax, multivariate hypergeometric, multinomial, negative hypergeometric, non-standard beta, normal mixture, Poisson mixture, Pareto, power, reparametrized beta, Rayleigh, shifted Gompertz, Skellam, slash, triangular, truncated binomial, truncated normal, truncated Poisson, Tukey lambda, Wald, zero-inflated binomial, zero-inflated negative binomial, zero-inflated Poisson.

Density, distribution function, quantile function and random generation for a number of univariate and multivariate distributions.

Details

This package follows naming convention that is consistent with base R, where density (or probability mass) functions, distribution functions, quantile functions and random generation functions names are followed by d*, p*, q*, and r* prefixes.

Behaviour of the functions is consistent with base R, where for not valid parameters values NaN's are returned, while for values beyond function support 0's are returned (e.g. for non-integers in discrete distributions, or for negative values in functions with non-negative support).

All the functions vectorized and coded in C++ using Rcpp.

Author(s)

Maintainer: Sigbert Klinke sigbert@wiwi.hu-berlin.de

Authors:

Tymoteusz Wolodzko twolodzko+extradistr@gmail.com

See Also

Useful links:

https://github.com/sigbertklinke/extraDistr

[Package extraDistr version 1.10.0.2 Index]